backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.5
fastembed==0.4.2
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
jiter==0.7.1
joblib==1.4.2
llama-index-core==0.12.1
llama-index-embeddings-fastembed==0.3.0
llama-index-llms-groq==0.3.0
llama-index-llms-openai==0.3.1
llama-index-llms-openai-like==0.3.0
llama-index-vector-stores-qdrant==0.4.0
llama-parse==0.5.15
loguru==0.7.2
marshmallow==3.23.1
mmh3==4.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.19.2
openai==1.55.0
packaging==24.2
pillow==10.4.0
portalocker==2.10.1
propcache==0.2.0
protobuf==5.28.3
py_rust_stemmers==0.1.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
qdrant-client==1.12.1
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.3
tenacity==8.5.0
tiktoken==0.8.0
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
wrapt==1.17.0
yarl==1.18.0
